Allow less potential CSS
Basically, if mimeType is failure we let it through because cross-origin CSS is allowed in that case. However, it's not allowed when nosniff is false per Fetch today. Additionally, Chrome and Safari ignore CSS without an ok status.

This means that when nosniff is true and mimeType is failure, we only need to allow sniffed media through.

Fixes #9.